requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==24.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
